+++ heat/core/tests/test_manipulations.py
@@ -371,13 +371,13 @@ class TestManipulations(TestCase):
         data = torch.arange(size * 2, device=self.device.torch_device)
         a = ht.array(data)
         res = ht.diag(a)
-        self.assertTrue(torch.equal(res._DNDarray__array, torch.diag(data)))
+        self.assertTrue(torch.equal(res.larray, torch.diag(data)))
 
         res = ht.diag(a, offset=size)
-        self.assertTrue(torch.equal(res._DNDarray__array, torch.diag(data, diagonal=size)))
+        self.assertTrue(torch.equal(res.larray, torch.diag(data, diagonal=size)))
 
         res = ht.diag(a, offset=-size)
-        self.assertTrue(torch.equal(res._DNDarray__array, torch.diag(data, diagonal=-size)))
+        self.assertTrue(torch.equal(res.larray, torch.diag(data, diagonal=-size)))
 
         a = ht.array(data, split=0)
         res = ht.diag(a)
@@ -386,7 +386,7 @@ class TestManipulations(TestCase):
         self.assertEqual(res.lshape[res.split], 2)
         exp = torch.diag(data)
         for i in range(rank * 2, (rank + 1) * 2):
-            self.assertTrue(torch.equal(res[i, i]._DNDarray__array, exp[i, i]))
+            self.assertTrue(torch.equal(res[i, i].larray, exp[i, i]))
 
         res = ht.diag(a, offset=size)
         self.assertEqual(res.split, a.split)
@@ -394,7 +394,7 @@ class TestManipulations(TestCase):
         self.assertEqual(res.lshape[res.split], 3)
         exp = torch.diag(data, diagonal=size)
         for i in range(rank * 3, min((rank + 1) * 3, a.shape[0])):
-            self.assertTrue(torch.equal(res[i, i + size]._DNDarray__array, exp[i, i + size]))
+            self.assertTrue(torch.equal(res[i, i + size].larray, exp[i, i + size]))
 
         res = ht.diag(a, offset=-size)
         self.assertEqual(res.split, a.split)
@@ -402,7 +402,7 @@ class TestManipulations(TestCase):
         self.assertEqual(res.lshape[res.split], 3)
         exp = torch.diag(data, diagonal=-size)
         for i in range(max(size, rank * 3), (rank + 1) * 3):
-            self.assertTrue(torch.equal(res[i, i - size]._DNDarray__array, exp[i, i - size]))
+            self.assertTrue(torch.equal(res[i, i - size].larray, exp[i, i - size]))
 
         self.assertTrue(ht.equal(ht.diag(ht.diag(a)), a))
 
@@ -433,7 +433,7 @@ class TestManipulations(TestCase):
         res = ht.diag(a)
         self.assertTrue(
             torch.equal(
-                res[rank, rank]._DNDarray__array,
+                res[rank, rank].larray,
                 torch.tensor(1, dtype=torch.int32, device=self.device.torch_device),
             )
         )
@@ -462,14 +462,14 @@ class TestManipulations(TestCase):
         a = ht.array(data)
         res = ht.diagonal(a)
         self.assertTrue(
-            torch.equal(res._DNDarray__array, torch.arange(size, device=self.device.torch_device))
+            torch.equal(res.larray, torch.arange(size, device=self.device.torch_device))
         )
         self.assertEqual(res.split, None)
 
         a = ht.array(data, split=0)
         res = ht.diagonal(a)
         self.assertTrue(
-            torch.equal(res._DNDarray__array, torch.tensor([rank], device=self.device.torch_device))
+            torch.equal(res.larray, torch.tensor([rank], device=self.device.torch_device))
         )
         self.assertEqual(res.split, 0)
 
@@ -479,7 +479,7 @@ class TestManipulations(TestCase):
 
         res = ht.diagonal(a)
         self.assertTrue(
-            torch.equal(res._DNDarray__array, torch.tensor([rank], device=self.device.torch_device))
+            torch.equal(res.larray, torch.tensor([rank], device=self.device.torch_device))
         )
         self.assertEqual(res.split, 0)
 
@@ -495,49 +495,39 @@ class TestManipulations(TestCase):
         a = ht.array(data)
         res = ht.diagonal(a, offset=0)
         self.assertTrue(
-            torch.equal(
-                res._DNDarray__array, torch.arange(size + 1, device=self.device.torch_device)
-            )
+            torch.equal(res.larray, torch.arange(size + 1, device=self.device.torch_device))
         )
         res = ht.diagonal(a, offset=1)
         self.assertTrue(
-            torch.equal(
-                res._DNDarray__array, torch.arange(1, size + 1, device=self.device.torch_device)
-            )
+            torch.equal(res.larray, torch.arange(1, size + 1, device=self.device.torch_device))
         )
         res = ht.diagonal(a, offset=-1)
         self.assertTrue(
-            torch.equal(
-                res._DNDarray__array, torch.arange(0, size, device=self.device.torch_device)
-            )
+            torch.equal(res.larray, torch.arange(0, size, device=self.device.torch_device))
         )
 
         a = ht.array(data, split=0)
         res = ht.diagonal(a, offset=1)
         res.balance_()
         self.assertTrue(
-            torch.equal(
-                res._DNDarray__array, torch.tensor([rank + 1], device=self.device.torch_device)
-            )
+            torch.equal(res.larray, torch.tensor([rank + 1], device=self.device.torch_device))
         )
         res = ht.diagonal(a, offset=-1)
         res.balance_()
         self.assertTrue(
-            torch.equal(res._DNDarray__array, torch.tensor([rank], device=self.device.torch_device))
+            torch.equal(res.larray, torch.tensor([rank], device=self.device.torch_device))
         )
 
         a = ht.array(data, split=1)
         res = ht.diagonal(a, offset=1)
         res.balance_()
         self.assertTrue(
-            torch.equal(
-                res._DNDarray__array, torch.tensor([rank + 1], device=self.device.torch_device)
-            )
+            torch.equal(res.larray, torch.tensor([rank + 1], device=self.device.torch_device))
         )
         res = ht.diagonal(a, offset=-1)
         res.balance_()
         self.assertTrue(
-            torch.equal(res._DNDarray__array, torch.tensor([rank], device=self.device.torch_device))
+            torch.equal(res.larray, torch.tensor([rank], device=self.device.torch_device))
         )
 
         data = (
@@ -549,15 +539,12 @@ class TestManipulations(TestCase):
         res = ht.diagonal(a, offset=10)
         self.assertTrue(
             torch.equal(
-                res._DNDarray__array,
-                torch.arange(10, 10 + size * 2, device=self.device.torch_device),
+                res.larray, torch.arange(10, 10 + size * 2, device=self.device.torch_device)
             )
         )
         res = ht.diagonal(a, offset=-10)
         self.assertTrue(
-            torch.equal(
-                res._DNDarray__array, torch.arange(0, size * 2, device=self.device.torch_device)
-            )
+            torch.equal(res.larray, torch.arange(0, size * 2, device=self.device.torch_device))
         )
 
         a = ht.array(data, split=0)
@@ -565,7 +552,7 @@ class TestManipulations(TestCase):
         res.balance_()
         self.assertTrue(
             torch.equal(
-                res._DNDarray__array,
+                res.larray,
                 torch.tensor([10 + rank * 2, 11 + rank * 2], device=self.device.torch_device),
             )
         )
@@ -573,8 +560,7 @@ class TestManipulations(TestCase):
         res.balance_()
         self.assertTrue(
             torch.equal(
-                res._DNDarray__array,
-                torch.tensor([rank * 2, 1 + rank * 2], device=self.device.torch_device),
+                res.larray, torch.tensor([rank * 2, 1 + rank * 2], device=self.device.torch_device)
             )
         )
 
@@ -583,7 +569,7 @@ class TestManipulations(TestCase):
         res.balance_()
         self.assertTrue(
             torch.equal(
-                res._DNDarray__array,
+                res.larray,
                 torch.tensor([10 + rank * 2, 11 + rank * 2], device=self.device.torch_device),
             )
         )
@@ -591,8 +577,7 @@ class TestManipulations(TestCase):
         res.balance_()
         self.assertTrue(
             torch.equal(
-                res._DNDarray__array,
-                torch.tensor([rank * 2, 1 + rank * 2], device=self.device.torch_device),
+                res.larray, torch.tensor([rank * 2, 1 + rank * 2], device=self.device.torch_device)
             )
         )
 
@@ -605,7 +590,7 @@ class TestManipulations(TestCase):
         res = ht.diagonal(a)
         self.assertTrue(
             torch.equal(
-                res._DNDarray__array,
+                res.larray,
                 torch.arange(size + 1, device=self.device.torch_device)
                 .repeat(size + 1)
                 .reshape(size + 1, size + 1)
@@ -615,7 +600,7 @@ class TestManipulations(TestCase):
         res = ht.diagonal(a, offset=1)
         self.assertTrue(
             torch.equal(
-                res._DNDarray__array,
+                res.larray,
                 torch.arange(size + 1, device=self.device.torch_device)
                 .repeat(size)
                 .reshape(size, size + 1)
@@ -625,7 +610,7 @@ class TestManipulations(TestCase):
         res = ht.diagonal(a, offset=-1)
         self.assertTrue(
             torch.equal(
-                res._DNDarray__array,
+                res.larray,
                 torch.arange(size + 1, device=self.device.torch_device)
                 .repeat(size)
                 .reshape(size, size + 1)
@@ -636,7 +621,7 @@ class TestManipulations(TestCase):
         res = ht.diagonal(a, dim1=1, dim2=2)
         self.assertTrue(
             torch.equal(
-                res._DNDarray__array,
+                res.larray,
                 torch.arange(size + 1, device=self.device.torch_device)
                 .repeat(size + 1)
                 .reshape(size + 1, size + 1),
@@ -645,7 +630,7 @@ class TestManipulations(TestCase):
         res = ht.diagonal(a, offset=1, dim1=1, dim2=2)
         self.assertTrue(
             torch.equal(
-                res._DNDarray__array,
+                res.larray,
                 torch.arange(1, size + 1, device=self.device.torch_device)
                 .repeat(size + 1)
                 .reshape(size + 1, size),
@@ -654,7 +639,7 @@ class TestManipulations(TestCase):
         res = ht.diagonal(a, offset=-1, dim1=1, dim2=2)
         self.assertTrue(
             torch.equal(
-                res._DNDarray__array,
+                res.larray,
                 torch.arange(size, device=self.device.torch_device)
                 .repeat(size + 1)
                 .reshape(size + 1, size),
@@ -664,7 +649,7 @@ class TestManipulations(TestCase):
         res = ht.diagonal(a, dim1=0, dim2=2)
         self.assertTrue(
             torch.equal(
-                res._DNDarray__array,
+                res.larray,
                 torch.arange(size + 1, device=self.device.torch_device)
                 .repeat(size + 1)
                 .reshape(size + 1, size + 1),
@@ -676,7 +661,7 @@ class TestManipulations(TestCase):
         res.balance_()
         self.assertTrue(
             torch.equal(
-                res._DNDarray__array,
+                res.larray,
                 torch.arange(size + 1, device=self.device.torch_device).reshape(size + 1, 1),
             )
         )
@@ -686,7 +671,7 @@ class TestManipulations(TestCase):
         res.balance_()
         self.assertTrue(
             torch.equal(
-                res._DNDarray__array,
+                res.larray,
                 torch.arange(size + 1, device=self.device.torch_device).reshape(size + 1, 1),
             )
         )
@@ -696,7 +681,7 @@ class TestManipulations(TestCase):
         res.balance_()
         self.assertTrue(
             torch.equal(
-                res._DNDarray__array,
+                res.larray,
                 torch.empty((size + 1, 0), dtype=torch.int64, device=self.device.torch_device),
             )
         )
@@ -1549,21 +1534,21 @@ class TestManipulations(TestCase):
         data = ht.array(tensor, split=None)
         result, result_indices = ht.sort(data, axis=0, descending=True)
         expected, exp_indices = torch.sort(tensor, dim=0, descending=True)
-        self.assertTrue(torch.equal(result._DNDarray__array, expected))
-        self.assertTrue(torch.equal(result_indices._DNDarray__array, exp_indices.int()))
+        self.assertTrue(torch.equal(result.larray, expected))
+        self.assertTrue(torch.equal(result_indices.larray, exp_indices.int()))
 
         result, result_indices = ht.sort(data, axis=1, descending=True)
         expected, exp_indices = torch.sort(tensor, dim=1, descending=True)
-        self.assertTrue(torch.equal(result._DNDarray__array, expected))
-        self.assertTrue(torch.equal(result_indices._DNDarray__array, exp_indices.int()))
+        self.assertTrue(torch.equal(result.larray, expected))
+        self.assertTrue(torch.equal(result_indices.larray, exp_indices.int()))
 
         data = ht.array(tensor, split=0)
 
         exp_axis_zero = torch.arange(size, device=self.device.torch_device).reshape(1, size)
         exp_indices = torch.tensor([[rank] * size], device=self.device.torch_device)
         result, result_indices = ht.sort(data, descending=True, axis=0)
-        self.assertTrue(torch.equal(result._DNDarray__array, exp_axis_zero))
-        self.assertTrue(torch.equal(result_indices._DNDarray__array, exp_indices.int()))
+        self.assertTrue(torch.equal(result.larray, exp_axis_zero))
+        self.assertTrue(torch.equal(result_indices.larray, exp_indices.int()))
 
         exp_axis_one, exp_indices = (
             torch.arange(size, device=self.device.torch_device)
@@ -1571,8 +1556,8 @@ class TestManipulations(TestCase):
             .sort(dim=1, descending=True)
         )
         result, result_indices = ht.sort(data, descending=True, axis=1)
-        self.assertTrue(torch.equal(result._DNDarray__array, exp_axis_one))
-        self.assertTrue(torch.equal(result_indices._DNDarray__array, exp_indices.int()))
+        self.assertTrue(torch.equal(result.larray, exp_axis_one))
+        self.assertTrue(torch.equal(result_indices.larray, exp_indices.int()))
 
         result1 = ht.sort(data, axis=1, descending=True)
         result2 = ht.sort(data, descending=True)
@@ -1588,10 +1573,10 @@ class TestManipulations(TestCase):
             size, dtype=torch.int64, device=self.device.torch_device
         ).reshape(size, 1)
         result, result_indices = ht.sort(data, axis=0, descending=True)
-        self.assertTrue(torch.equal(result._DNDarray__array, exp_axis_zero))
+        self.assertTrue(torch.equal(result.larray, exp_axis_zero))
         # comparison value is only true on CPU
-        if result_indices._DNDarray__array.is_cuda is False:
-            self.assertTrue(torch.equal(result_indices._DNDarray__array, indices_axis_zero.int()))
+        if result_indices.larray.is_cuda is False:
+            self.assertTrue(torch.equal(result_indices.larray, indices_axis_zero.int()))
 
         exp_axis_one = (
             torch.tensor(size - rank - 1, device=self.device.torch_device)
@@ -1599,8 +1584,8 @@ class TestManipulations(TestCase):
             .reshape(size, 1)
         )
         result, result_indices = ht.sort(data, descending=True, axis=1)
-        self.assertTrue(torch.equal(result._DNDarray__array, exp_axis_one))
-        self.assertTrue(torch.equal(result_indices._DNDarray__array, exp_axis_one.int()))
+        self.assertTrue(torch.equal(result.larray, exp_axis_one))
+        self.assertTrue(torch.equal(result_indices.larray, exp_axis_one.int()))
 
         tensor = torch.tensor(
             [
@@ -1626,8 +1611,8 @@ class TestManipulations(TestCase):
                 [[0, 2, 2], [3, 0, 0]], dtype=torch.int32, device=self.device.torch_device
             )
         result, result_indices = ht.sort(data, axis=0)
-        first = result[0]._DNDarray__array
-        first_indices = result_indices[0]._DNDarray__array
+        first = result[0].larray
+        first_indices = result_indices[0].larray
         if rank == 0:
             self.assertTrue(torch.equal(first, exp_axis_zero))
             self.assertTrue(torch.equal(first_indices, indices_axis_zero))
@@ -1638,8 +1623,8 @@ class TestManipulations(TestCase):
             [[0, 1, 1]], dtype=torch.int32, device=self.device.torch_device
         )
         result, result_indices = ht.sort(data, axis=1)
-        first = result[0]._DNDarray__array[:1]
-        first_indices = result_indices[0]._DNDarray__array[:1]
+        first = result[0].larray[:1]
+        first_indices = result_indices[0].larray[:1]
         if rank == 0:
             self.assertTrue(torch.equal(first, exp_axis_one))
             self.assertTrue(torch.equal(first_indices, indices_axis_one))
@@ -1650,8 +1635,8 @@ class TestManipulations(TestCase):
             [[0], [1]], dtype=torch.int32, device=self.device.torch_device
         )
         result, result_indices = ht.sort(data, axis=2)
-        first = result[0]._DNDarray__array[:, :1]
-        first_indices = result_indices[0]._DNDarray__array[:, :1]
+        first = result[0].larray[:, :1]
+        first_indices = result_indices[0].larray[:, :1]
         if rank == 0:
             self.assertTrue(torch.equal(first, exp_axis_two))
             self.assertTrue(torch.equal(first_indices, indices_axis_two))
@@ -1674,11 +1659,7 @@ class TestManipulations(TestCase):
         for i, c in enumerate(counts):
             for idx in range(c - 1):
                 if rank == i:
-                    self.assertTrue(
-                        torch.lt(
-                            result._DNDarray__array[idx], result._DNDarray__array[idx + 1]
-                        ).all()
-                    )
+                    self.assertTrue(torch.lt(result.larray[idx], result.larray[idx + 1]).all())
 
     def test_resplit(self):
         if ht.MPI_WORLD.size > 1:
@@ -1746,15 +1727,11 @@ class TestManipulations(TestCase):
             local_shape = (1, N + 1, 2 * N)
             local_tensor = reference_tensor[ht.MPI_WORLD.rank, :, :]
             self.assertEqual(resplit_tensor.lshape, local_shape)
-            self.assertTrue(
-                (resplit_tensor._DNDarray__array == local_tensor._DNDarray__array).all()
-            )
+            self.assertTrue((resplit_tensor.larray == local_tensor.larray).all())
 
             # unsplit
             unsplit_tensor = ht.resplit(resplit_tensor, axis=None)
-            self.assertTrue(
-                (unsplit_tensor._DNDarray__array == reference_tensor._DNDarray__array).all()
-            )
+            self.assertTrue((unsplit_tensor.larray == reference_tensor.larray).all())
 
             # split along axis = 1
             resplit_tensor = ht.resplit(unsplit_tensor, axis=1)
@@ -1766,15 +1743,11 @@ class TestManipulations(TestCase):
                 local_tensor = reference_tensor[:, ht.MPI_WORLD.rank + 1 : ht.MPI_WORLD.rank + 2, :]
 
             self.assertEqual(resplit_tensor.lshape, local_shape)
-            self.assertTrue(
-                (resplit_tensor._DNDarray__array == local_tensor._DNDarray__array).all()
-            )
+            self.assertTrue((resplit_tensor.larray == local_tensor.larray).all())
 
             # unsplit
             unsplit_tensor = ht.resplit(resplit_tensor, axis=None)
-            self.assertTrue(
-                (unsplit_tensor._DNDarray__array == reference_tensor._DNDarray__array).all()
-            )
+            self.assertTrue((unsplit_tensor.larray == reference_tensor.larray).all())
 
             # split along axis = 2
             resplit_tensor = ht.resplit(unsplit_tensor, axis=2)
@@ -1782,9 +1755,7 @@ class TestManipulations(TestCase):
             local_tensor = reference_tensor[:, :, 2 * ht.MPI_WORLD.rank : 2 * ht.MPI_WORLD.rank + 2]
 
             self.assertEqual(resplit_tensor.lshape, local_shape)
-            self.assertTrue(
-                (resplit_tensor._DNDarray__array == local_tensor._DNDarray__array).all()
-            )
+            self.assertTrue((resplit_tensor.larray == local_tensor.larray).all())
 
             # order tests for resplit
             for dims in range(3, 5):
@@ -1811,48 +1782,48 @@ class TestManipulations(TestCase):
         result = ht.squeeze(data)
         self.assertIsInstance(result, ht.DNDarray)
         self.assertEqual(result.dtype, ht.float32)
-        self.assertEqual(result._DNDarray__array.dtype, torch.float32)
+        self.assertEqual(result.larray.dtype, torch.float32)
         self.assertEqual(result.shape, (4, 5))
         self.assertEqual(result.lshape, (4, 5))
         self.assertEqual(result.split, None)
-        self.assertTrue((result._DNDarray__array == data._DNDarray__array.squeeze()).all())
+        self.assertTrue((result.larray == data.larray.squeeze()).all())
 
         # 4D local tensor, major axis
         result = ht.squeeze(data, axis=0)
         self.assertIsInstance(result, ht.DNDarray)
         self.assertEqual(result.dtype, ht.float32)
-        self.assertEqual(result._DNDarray__array.dtype, torch.float32)
+        self.assertEqual(result.larray.dtype, torch.float32)
         self.assertEqual(result.shape, (4, 5, 1))
         self.assertEqual(result.lshape, (4, 5, 1))
         self.assertEqual(result.split, None)
-        self.assertTrue((result._DNDarray__array == data._DNDarray__array.squeeze(0)).all())
+        self.assertTrue((result.larray == data.larray.squeeze(0)).all())
 
         # 4D local tensor, minor axis
         result = ht.squeeze(data, axis=-1)
         self.assertIsInstance(result, ht.DNDarray)
         self.assertEqual(result.dtype, ht.float32)
-        self.assertEqual(result._DNDarray__array.dtype, torch.float32)
+        self.assertEqual(result.larray.dtype, torch.float32)
         self.assertEqual(result.shape, (1, 4, 5))
         self.assertEqual(result.lshape, (1, 4, 5))
         self.assertEqual(result.split, None)
-        self.assertTrue((result._DNDarray__array == data._DNDarray__array.squeeze(-1)).all())
+        self.assertTrue((result.larray == data.larray.squeeze(-1)).all())
 
         # 4D local tensor, tuple axis
         result = data.squeeze(axis=(0, -1))
         self.assertIsInstance(result, ht.DNDarray)
         self.assertEqual(result.dtype, ht.float32)
-        self.assertEqual(result._DNDarray__array.dtype, torch.float32)
+        self.assertEqual(result.larray.dtype, torch.float32)
         self.assertEqual(result.shape, (4, 5))
         self.assertEqual(result.lshape, (4, 5))
         self.assertEqual(result.split, None)
-        self.assertTrue((result._DNDarray__array == data._DNDarray__array.squeeze()).all())
+        self.assertTrue((result.larray == data.larray.squeeze()).all())
 
         # 4D split tensor, along the axis
         data = ht.array(ht.random.randn(1, 4, 5, 1), split=1)
         result = ht.squeeze(data, axis=-1)
         self.assertIsInstance(result, ht.DNDarray)
         self.assertEqual(result.dtype, ht.float32)
-        self.assertEqual(result._DNDarray__array.dtype, torch.float32)
+        self.assertEqual(result.larray.dtype, torch.float32)
         self.assertEqual(result.shape, (1, 4, 5))
         self.assertEqual(result.split, 1)
 
@@ -1861,7 +1832,7 @@ class TestManipulations(TestCase):
         result = ht.squeeze(data, axis=1)
         self.assertIsInstance(result, ht.DNDarray)
         self.assertEqual(result.dtype, ht.float32)
-        self.assertEqual(result._DNDarray__array.dtype, torch.float32)
+        self.assertEqual(result.larray.dtype, torch.float32)
         self.assertEqual(result.shape, (3, 5, 6))
         self.assertEqual(result.split, None)
 
@@ -1870,7 +1841,7 @@ class TestManipulations(TestCase):
         result = ht.squeeze(data, axis=-1)
         self.assertIsInstance(result, ht.DNDarray)
         self.assertEqual(result.dtype, ht.float32)
-        self.assertEqual(result._DNDarray__array.dtype, torch.float32)
+        self.assertEqual(result.larray.dtype, torch.float32)
         self.assertEqual(result.shape, (3, 6, 5))
         self.assertEqual(result.split, None)
 
@@ -1881,7 +1852,7 @@ class TestManipulations(TestCase):
         result = ht.squeeze(data, axis=0)
         self.assertIsInstance(result, ht.DNDarray)
         self.assertEqual(result.dtype, ht.float32)
-        self.assertEqual(result._DNDarray__array.dtype, torch.float32)
+        self.assertEqual(result.larray.dtype, torch.float32)
         self.assertEqual(result.shape, (size * 2, size))
         self.assertEqual(result.lshape, (2, size))
         self.assertEqual(result.split, 0)
@@ -2005,18 +1976,18 @@ class TestManipulations(TestCase):
         exp_zero_indcs = ht.array(
             [[size - 1, size - 2] for i in range(size)], dtype=ht.int64, split=0
         )
-        self.assertTrue((res._DNDarray__array == exp_zero._DNDarray__array).all())
-        self.assertTrue((indcs._DNDarray__array == exp_zero._DNDarray__array).all())
-        self.assertTrue(indcs._DNDarray__array.dtype == exp_zero_indcs._DNDarray__array.dtype)
+        self.assertTrue((res.larray == exp_zero.larray).all())
+        self.assertTrue((indcs.larray == exp_zero.larray).all())
+        self.assertTrue(indcs.larray.dtype == exp_zero_indcs.larray.dtype)
 
         res, indcs = ht.topk(split_one, 2, sorted=True)
         exp_one = ht.array([[size - 1, size - 2] for i in range(size)], dtype=ht.int32, split=1)
         exp_one_indcs = ht.array(
             [[size - 1, size - 2] for i in range(size)], dtype=ht.int64, split=1
         )
-        self.assertTrue((res._DNDarray__array == exp_one._DNDarray__array).all())
-        self.assertTrue((indcs._DNDarray__array == exp_one_indcs._DNDarray__array).all())
-        self.assertTrue(indcs._DNDarray__array.dtype == exp_one_indcs._DNDarray__array.dtype)
+        self.assertTrue((res.larray == exp_one.larray).all())
+        self.assertTrue((indcs.larray == exp_one_indcs.larray).all())
+        self.assertTrue(indcs.larray.dtype == exp_one_indcs.larray.dtype)
 
         torch_array = torch.arange(size, dtype=torch.float64).expand(size, size)
         split_zero = ht.array(torch_array, split=0)
@@ -2027,38 +1998,38 @@ class TestManipulations(TestCase):
         exp_zero_indcs = ht.array(
             [[size - 1, size - 2] for i in range(size)], dtype=ht.int64, split=0
         )
-        self.assertTrue((res._DNDarray__array == exp_zero._DNDarray__array).all())
-        self.assertTrue((indcs._DNDarray__array == exp_zero_indcs._DNDarray__array).all())
-        self.assertTrue(indcs._DNDarray__array.dtype == exp_zero_indcs._DNDarray__array.dtype)
+        self.assertTrue((res.larray == exp_zero.larray).all())
+        self.assertTrue((indcs.larray == exp_zero_indcs.larray).all())
+        self.assertTrue(indcs.larray.dtype == exp_zero_indcs.larray.dtype)
 
         res, indcs = ht.topk(split_one, 2, sorted=True)
         exp_one = ht.array([[size - 1, size - 2] for i in range(size)], dtype=ht.float64, split=1)
         exp_one_indcs = ht.array(
             [[size - 1, size - 2] for i in range(size)], dtype=ht.int64, split=1
         )
-        self.assertTrue((res._DNDarray__array == exp_one._DNDarray__array).all())
-        self.assertTrue((indcs._DNDarray__array == exp_one_indcs._DNDarray__array).all())
-        self.assertTrue(indcs._DNDarray__array.dtype == exp_one_indcs._DNDarray__array.dtype)
+        self.assertTrue((res.larray == exp_one.larray).all())
+        self.assertTrue((indcs.larray == exp_one_indcs.larray).all())
+        self.assertTrue(indcs.larray.dtype == exp_one_indcs.larray.dtype)
 
         res, indcs = ht.topk(split_zero, 2, sorted=True, largest=False)
         exp_zero = ht.array([[0, 1] for i in range(size)], dtype=ht.int32, split=0)
         exp_zero_indcs = ht.array([[0, 1] for i in range(size)], dtype=ht.int64, split=0)
-        self.assertTrue((res._DNDarray__array == exp_zero._DNDarray__array).all())
-        self.assertTrue((indcs._DNDarray__array == exp_zero._DNDarray__array).all())
-        self.assertTrue(indcs._DNDarray__array.dtype == exp_zero_indcs._DNDarray__array.dtype)
+        self.assertTrue((res.larray == exp_zero.larray).all())
+        self.assertTrue((indcs.larray == exp_zero.larray).all())
+        self.assertTrue(indcs.larray.dtype == exp_zero_indcs.larray.dtype)
 
         exp_zero = ht.array([[0, 1] for i in range(size)], dtype=ht.int32, split=0)
         exp_zero_indcs = ht.array([[0, 1] for i in range(size)], dtype=ht.int64, split=0)
         out = (ht.empty_like(exp_zero), ht.empty_like(exp_zero_indcs))
         res, indcs = ht.topk(split_zero, 2, sorted=True, largest=False, out=out)
 
-        self.assertTrue((res._DNDarray__array == exp_zero._DNDarray__array).all())
-        self.assertTrue((indcs._DNDarray__array == exp_zero._DNDarray__array).all())
-        self.assertTrue(indcs._DNDarray__array.dtype == exp_zero_indcs._DNDarray__array.dtype)
+        self.assertTrue((res.larray == exp_zero.larray).all())
+        self.assertTrue((indcs.larray == exp_zero.larray).all())
+        self.assertTrue(indcs.larray.dtype == exp_zero_indcs.larray.dtype)
 
-        self.assertTrue((out[0]._DNDarray__array == exp_zero._DNDarray__array).all())
-        self.assertTrue((out[1]._DNDarray__array == exp_zero._DNDarray__array).all())
-        self.assertTrue(out[1]._DNDarray__array.dtype == exp_zero_indcs._DNDarray__array.dtype)
+        self.assertTrue((out[0].larray == exp_zero.larray).all())
+        self.assertTrue((out[1].larray == exp_zero.larray).all())
+        self.assertTrue(out[1].larray.dtype == exp_zero_indcs.larray.dtype)
 
     def test_unique(self):
         size = ht.MPI_WORLD.size
@@ -2070,30 +2041,30 @@ class TestManipulations(TestCase):
 
         exp_axis_none = ht.array([rank], dtype=ht.int32)
         res = split_zero.unique(sorted=True)
-        self.assertTrue((res._DNDarray__array == exp_axis_none._DNDarray__array).all())
+        self.assertTrue((res.larray == exp_axis_none.larray).all())
 
         exp_axis_zero = ht.arange(size, dtype=ht.int32).expand_dims(0)
         res = ht.unique(split_zero, sorted=True, axis=0)
-        self.assertTrue((res._DNDarray__array == exp_axis_zero._DNDarray__array).all())
+        self.assertTrue((res.larray == exp_axis_zero.larray).all())
 
         exp_axis_one = ht.array([rank], dtype=ht.int32).expand_dims(0)
         split_zero_transposed = ht.array(torch_array.transpose(0, 1), split=0)
         res = ht.unique(split_zero_transposed, sorted=False, axis=1)
-        self.assertTrue((res._DNDarray__array == exp_axis_one._DNDarray__array).all())
+        self.assertTrue((res.larray == exp_axis_one.larray).all())
 
         split_one = ht.array(torch_array, dtype=ht.int32, split=1)
 
         exp_axis_none = ht.arange(size, dtype=ht.int32)
         res = ht.unique(split_one, sorted=True)
-        self.assertTrue((res._DNDarray__array == exp_axis_none._DNDarray__array).all())
+        self.assertTrue((res.larray == exp_axis_none.larray).all())
 
         exp_axis_zero = ht.array([rank], dtype=ht.int32).expand_dims(0)
         res = ht.unique(split_one, sorted=False, axis=0)
-        self.assertTrue((res._DNDarray__array == exp_axis_zero._DNDarray__array).all())
+        self.assertTrue((res.larray == exp_axis_zero.larray).all())
 
         exp_axis_one = ht.array([rank] * size, dtype=ht.int32).expand_dims(1)
         res = ht.unique(split_one, sorted=True, axis=1)
-        self.assertTrue((res._DNDarray__array == exp_axis_one._DNDarray__array).all())
+        self.assertTrue((res.larray == exp_axis_one.larray).all())
 
         torch_array = torch.tensor(
             [[1, 2], [2, 3], [1, 2], [2, 3], [1, 2]],
@@ -2128,7 +2099,7 @@ class TestManipulations(TestCase):
         self.assertEqual(inv.split, None)
         self.assertEqual(inv.dtype, data_split_none.dtype)
         self.assertEqual(inv.device, data_split_none.device)
-        self.assertTrue(torch.equal(inv._DNDarray__array, exp_inv.int()))
+        self.assertTrue(torch.equal(inv.larray, exp_inv.int()))
 
         data_split_zero = ht.array(torch_array, split=0)
         res, inv = ht.unique(data_split_zero, return_inverse=True, sorted=True)